drowaudio-reverb: Fix working on higher sample rates
The tap delay buffer should be able to hold a few seconds of audio.
When it does not, it leads to out-of-bounds reads, which may lead
to segfaults or other glitches.

Following the infamous "640k ought to be enough for anybody" motto,
we increase the default tap delay buffer size to make it large enough
for any reasonable sample rate. This increases RAM usage of the plugin
by ~1Mb, which seems reasonable by modern day standards.